Restructure module & remove serde types from db crate
