+++ crates/katana/storage/db/Cargo.toml
@@ -16,9 +16,3 @@ page_size = "0.6.0"
 parking_lot.workspace = true
 serde.workspace = true
 thiserror.workspace = true
-
-# blockifier types serde compatiblity
-blockifier.workspace = true
-cairo-vm.workspace = true
-serde_json.workspace = true
-starknet_api.workspace = true
