--- crates/katana/storage/db/src/mdbx/models/mod.rs
@@ -1,2 +1,0 @@
-pub mod block;
-pub mod contract;
