--- crates/katana/storage/db/src/serde/mod.rs
@@ -1,3 +1,0 @@
-//! Serde support for types that are not de/serializable out of the box.
-
-pub mod blockifier;
